Fix endless spin due to competition over Sercoms
SPI neopixel would compete with i2c over the same sercoms. The correct patch here would be to use SERCOM4 for neopixel, but for now just use the .h version of the neopixel driver.

The minimum size required to use spi neopixel has increased to 24 bytes, which is 8 rgb values.