Replace sineTone lookup table with approximation
Instead of a 512-byte lookup table, use a 5th order polynomial
approximation.

The maximum error is 7/32767 = 0.02% which should be good enough
for sound generation purposes, especially on target devices with
small non-HiFi speakers.

For the record, the approximation uses an odd 5th-order polynomial,
sin(x * pi/2) = cx^5 + bx^3 + ax, mapped from 0..pi/2 to 0..1, with the
constraint that f(1)=1. Solving for that results in a = 1 - b - c.

I used a least-squares fit in Emacs Calc to calculate the parameters
a, b, and c:

  [
  # Create a 21-element vector [0, 0.05, ..., 0.95, 1]
  v x 21 RET
  1 -
  20 /
  # Duplicate it, map each element to the corresponding sine value
  RET
  V M ' sin(x * 90) RET RET
  ]
  # Least squares fit the custom function.
  a F ' (1 - b - c) x + b x^3 + c x^5 RET RET

=> 1.54543280892 x - 0.598500841146 x^3 + 0.0530680322224 x^5

(See https://www.coranac.com/2009/07/sines/ for background info.
I used different constraints, the result may be a bit less accurate
but seemed more straightforward to calculate.)