Update platform.h to include new pins for Adafruit Grand Central M4
on behalf of @lalauria2000. Part 2 of 2.

See: microsoft/pxt-maker#324

May need to find a better way to move to the 128 pin packages.  I just made it so that if the board is SAMD51, it will always default to 128 pins.  Not sure if this is 100% accurate since a lot of the variants do not have 128 pins.