Respect Transparency for Screen Edge Collisions
This change switches "Stay in Screen" and "Bounce on Wall" checks to use the sprite's Hitbox instead of the left, right, top, and bottom fields. Left, right, top, and bottom do not account for transparency around the sprite, so a sprite with a transparent border would look like it could never actually reach the edge. By using HitBox, we ensure the sprite will not stop/bounce until a non-transparent section of the sprite intersects with the edge.

I've also moved this check into the PhysicsEngine, which seems like a slightly more natural place for it.