Fix radio extension's "more" toolbox display
The raiseEvent block was referencing block ids control_event_source_id and control_event_value_id which don't exist in common packages. I didn't remove the code completely in case anyone already has this block in their project, but I don't think we need to show it. In live, it's already basically impossible to access because it messes up the toolbox load.

One unknown is if this is needed by micro:bit, which does actually have the block ids. If that's the case, correct fix may be to actually bring the missing block into common packages instead.